# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NDVI: восстановление ряда и аномалии",
    "version": "1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/meta": {
      "get": {
        "summary": "Meta",
        "description": "Всё, что фронту нужно знать до первого запроса.\n\nРамка региона здесь не украшение: модель обучена на Ростовской области,\nи за её пределами ответ будет уверенной чепухой. Ограничивай рисование.",
        "operationId": "meta_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Meta Meta Get"
                }
              }
            }
          }
        }
      }
    },
    "/predict": {
      "post": {
        "summary": "Predict",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Predict Predict Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyze": {
      "post": {
        "summary": "Analyze",
        "description": "Область на карте -> задача. Результат забирать через GET /analyze/{job_id}.\n\nВыгрузка из Earth Engine занимает около минуты: столько держать открытый\nHTTP-запрос нельзя, браузер и прокси его оборвут.",
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Analyze Analyze Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyze/{job_id}": {
      "get": {
        "summary": "Analyze Status",
        "operationId": "analyze_status_analyze__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Analyze Status Analyze  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "geometry": {
            "type": "object",
            "title": "Geometry",
            "description": "GeoJSON Polygon или MultiPolygon, WGS84"
          },
          "years": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Years",
            "description": "[первый, последний] или полный список",
            "default": [
              2016,
              2025
            ]
          },
          "crop_type": {
            "type": "string",
            "title": "Crop Type",
            "description": "из /meta -> crop_types",
            "default": "неизвестно"
          }
        },
        "type": "object",
        "required": [
          "geometry"
        ],
        "title": "AnalyzeRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PredictRequest": {
        "properties": {
          "rows": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Rows",
            "description": "строки в схеме исходных данных"
          },
          "with_anomalies": {
            "type": "boolean",
            "title": "With Anomalies",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "PredictRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}